backend/requirements.txt
# Legal AI System Dependencies
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
google-generativeai==0.8.3
cryptography==41.0.8
sentence-transformers==2.2.2
chromadb==0.4.24
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
requests==2.31.0
Werkzeug==3.0.1
pytest==7.4.3
regex==2023.10.3